[codeql] Widen 16 int-multiplication-before-cast sites to avoid overflow
CodeQL's cpp/integer-multiplication-cast-to-long flagged 16 new sites
where two narrower operands (unsigned int, or a uint32_t returned by
align()/ceilDiv()) are multiplied and the unsigned-int result is then
implicitly converted to a wider type -- size_t, int64_t, or a vector's
size_type -- for an allocation size, buffer offset, or element count.
On dimensions large enough to overflow 32 bits, the multiply wraps
before the widening happens, so the destination sees a truncated value
instead of the true count.

Each site now casts one operand to the destination width before the
multiply: nntrainer/layers/llm/tie_word_embedding.cpp (per-row weight
offset), nntrainer/tensor/float_tensor.cpp (broadcast stride offset),
nntrainer/tensor/int4_utils.cpp (padded scales buffer size), and the
row/column products test/unittest/unittest_nntrainer_cpu_backend.cpp
builds vectors and calls quantization helpers with.

+++ nntrainer/tensor/int4_utils.cpp
@@ -63,7 +63,8 @@ void Int4Utils::computeScales(const float *weights, const size_t rows_count,
   const auto last_group_size = columns_count % group_size;
   const auto padded_groups_per_row = ceilDiv(columns_count, group_size);
   const auto rows_count_pad = align(rows_count, ROW_BLOCK_SIZE);
-  scales.resize(rows_count_pad * padded_groups_per_row, 1.0f);
+  scales.resize(static_cast<size_t>(rows_count_pad) * padded_groups_per_row,
+                1.0f);
 
   for (size_t row_id = 0; row_id < rows_count; ++row_id) {
     const auto *weights_row = weights + (row_id * columns_count);
